remove duplication of server local file support tests
